Finised iteration 2: recipes page
Created a new directory called recipes so that I can store my favorite dish inside of. Then I created a html file called chowmein.html inside the recipes which is my favorite dish.
Then I went back to the index.html that we created in the first iteration and created a link that links to the chowmein recipe so theres a connection between the two pages